tools/video/requirements.txt
google-genai>=1.47.0
python-dotenv>=1.0.0
PyJWT>=2.0.0
requests>=2.28.0
